Upgrade Pinot version to 1.3.0 & move off of presto-pinot-driver
If implemented this will upgrade the Pinot API version to 1.3.0. It will also
move off of the shaded presto-pinot-driver dependency.
