+++ presto-elasticsearch/pom.xml
@@ -16,6 +16,7 @@
         <dep.elasticsearch.version>7.17.27</dep.elasticsearch.version>
         <dep.log4j.version>2.24.3</dep.log4j.version>
         <air.check.skip-modernizer>true</air.check.skip-modernizer>
+        <dep.lucene.version>8.11.3</dep.lucene.version>
     </properties>
 
     <dependencies>
